Add
Adição do exercício 3 e parte do exercício 4.

+++ Lista13/ex3.cpp
@@ -1,7 +1,7 @@
 #include<bits/stdc++.h>
 using namespace std;
 
-float x=0,w=0,fahn;
+int x=0,w=0,fahn;
 int main()
 {
     cin>> x >> w;
